First attempt at trying to move RouteGroupItem.Validate into sep package
