+++ dataclients/kubernetes/routegroups/testdata/validation/invalid-algorithm.log
