+++ dataclients/kubernetes/routegroups/testdata/validation/path-and-path-subtree-in-the-same-route.yaml
