+++ dataclients/kubernetes/routegroups/testdata/validation/route-with-empty-filter.yaml
