+++ dataclients/kubernetes/routegroups/testdata/validation/route-with-empty-method.yaml
