+++ dataclients/kubernetes/routegroups/testdata/validation/route-with-empty-predicate.yaml
